48 - Clarified "Törlés" button and added default confirmation for deleting race

+++ workspace/apps/frontend/src/app/components/pages/competition/competition-editor/competition-editor.component.ts
@@ -234,7 +234,10 @@ export class CompetitionEditorComponent implements OnInit, OnDestroy {
     if (this.id == null) {
       return;
     }
-    this.competitionService.deleteCompetition(this.id).subscribe(() => this.router.navigate(['competition']));
+
+    if (window.confirm('Biztosan törölni szeretnéd ezt a versenyt?')) {
+      this.competitionService.deleteCompetition(this.id).subscribe(() => this.router.navigate(['competition']));
+    }
   }
 
   addStudentRow() {
